Port program loading and tests to the 64-bit runtime

+++ src/programs/args/main.c
@@ -2,8 +2,8 @@
 
 uint32_t program_main(const ProgramApi* api, const ProgramInfo* info)
 {
-    uint32_t* argv;
-    uint32_t i;
+    uintptr_t* argv;
+    size_t i;
 
     (void)api;
 
@@ -11,7 +11,7 @@ uint32_t program_main(const ProgramApi* api, const ProgramInfo* info)
     program_write_u32(info->argc);
     program_write("\n", 1);
 
-    argv = (uint32_t*)(uintptr_t)info->argv_addr;
+    argv = (uintptr_t*)(uintptr_t)info->argv_addr;
     for (i = 0; i < info->argc; ++i) {
         program_write("argv[", 5);
         program_write_u32(i);
